pyproject.toml
[tool.poetry]
name = "letter-be"
version = "0.1.0"
description = ""
authors = ["gmyun1999 <dbsrbals26@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.0.6"
django-dotenv = "1.4.2"
PyJWT = "2.8.0"
pydantic = "2.7.4"
djangorestframework = "^3.15.2"
types-pymysql = "^1.1.0.20240524"
django-cors-headers = "^4.4.0"
arrow = "^1.3.0"
dacite = "^1.8.1"
httpx = "^0.27.2"
ulid-py = "^1.1.0"
bleach = "^6.1.0"
firebase-admin = "^6.5.0"
django-environ = "^0.11.2"
pymysql = "^1.1.1"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
